moved boostrap to live with the binary
